\name{dataR}
\alias{dataR}
\docType{data}
\title{dataR}
\description{
Dataset R for usage examples.
}
\usage{data("dataR")}
\format{
  A data frame with 87 observations on the following 125 variables.
  \describe{
    \item{\code{x1}}{a numeric vector}
    \item{\code{x2}}{a numeric vector}
    \item{\code{x3}}{a numeric vector}
    \item{\code{x4}}{a numeric vector}
    \item{\code{x5}}{a numeric vector}
    \item{\code{x6}}{a numeric vector}
    \item{\code{x7}}{a numeric vector}
    \item{\code{x8}}{a numeric vector}
    \item{\code{x9}}{a numeric vector}
    \item{\code{x10}}{a numeric vector}
    \item{\code{x11}}{a numeric vector}
    \item{\code{x12}}{a numeric vector}
    \item{\code{x13}}{a numeric vector}
    \item{\code{x14}}{a numeric vector}
    \item{\code{x15}}{a numeric vector}
    \item{\code{x16}}{a numeric vector}
    \item{\code{x17}}{a numeric vector}
    \item{\code{x18}}{a numeric vector}
    \item{\code{x19}}{a numeric vector}
    \item{\code{x20}}{a numeric vector}
    \item{\code{x21}}{a numeric vector}
    \item{\code{x22}}{a numeric vector}
    \item{\code{x23}}{a numeric vector}
    \item{\code{x24}}{a numeric vector}
    \item{\code{x25}}{a numeric vector}
    \item{\code{x26}}{a numeric vector}
    \item{\code{x27}}{a numeric vector}
    \item{\code{x28}}{a numeric vector}
    \item{\code{x29}}{a numeric vector}
    \item{\code{x30}}{a numeric vector}
    \item{\code{x31}}{a numeric vector}
    \item{\code{x32}}{a numeric vector}
    \item{\code{x33}}{a numeric vector}
    \item{\code{x34}}{a numeric vector}
    \item{\code{x35}}{a numeric vector}
    \item{\code{x36}}{a numeric vector}
    \item{\code{x37}}{a numeric vector}
    \item{\code{x38}}{a numeric vector}
    \item{\code{x39}}{a numeric vector}
    \item{\code{x40}}{a numeric vector}
    \item{\code{x41}}{a numeric vector}
    \item{\code{x42}}{a numeric vector}
    \item{\code{x43}}{a numeric vector}
    \item{\code{x44}}{a numeric vector}
    \item{\code{x45}}{a numeric vector}
    \item{\code{x46}}{a numeric vector}
    \item{\code{x47}}{a numeric vector}
    \item{\code{x48}}{a numeric vector}
    \item{\code{x49}}{a numeric vector}
    \item{\code{x50}}{a numeric vector}
    \item{\code{x51}}{a numeric vector}
    \item{\code{x52}}{a numeric vector}
    \item{\code{x53}}{a numeric vector}
    \item{\code{x54}}{a numeric vector}
    \item{\code{x55}}{a numeric vector}
    \item{\code{x56}}{a numeric vector}
    \item{\code{x57}}{a numeric vector}
    \item{\code{x58}}{a numeric vector}
    \item{\code{x59}}{a numeric vector}
    \item{\code{x60}}{a numeric vector}
    \item{\code{x61}}{a numeric vector}
    \item{\code{x62}}{a numeric vector}
    \item{\code{x63}}{a numeric vector}
    \item{\code{x64}}{a numeric vector}
    \item{\code{x65}}{a numeric vector}
    \item{\code{x66}}{a numeric vector}
    \item{\code{x67}}{a numeric vector}
    \item{\code{x68}}{a numeric vector}
    \item{\code{x69}}{a numeric vector}
    \item{\code{x70}}{a numeric vector}
    \item{\code{x71}}{a numeric vector}
    \item{\code{x72}}{a numeric vector}
    \item{\code{x73}}{a numeric vector}
    \item{\code{x74}}{a numeric vector}
    \item{\code{x75}}{a numeric vector}
    \item{\code{x76}}{a numeric vector}
    \item{\code{x77}}{a numeric vector}
    \item{\code{x78}}{a numeric vector}
    \item{\code{x79}}{a numeric vector}
    \item{\code{x80}}{a numeric vector}
    \item{\code{x81}}{a numeric vector}
    \item{\code{x82}}{a numeric vector}
    \item{\code{x83}}{a numeric vector}
    \item{\code{x84}}{a numeric vector}
    \item{\code{x85}}{a numeric vector}
    \item{\code{x86}}{a numeric vector}
    \item{\code{x87}}{a numeric vector}
    \item{\code{x88}}{a numeric vector}
    \item{\code{x89}}{a numeric vector}
    \item{\code{x90}}{a numeric vector}
    \item{\code{x91}}{a numeric vector}
    \item{\code{x92}}{a numeric vector}
    \item{\code{x93}}{a numeric vector}
    \item{\code{x94}}{a numeric vector}
    \item{\code{x95}}{a numeric vector}
    \item{\code{x96}}{a numeric vector}
    \item{\code{x97}}{a numeric vector}
    \item{\code{x98}}{a numeric vector}
    \item{\code{x99}}{a numeric vector}
    \item{\code{x100}}{a numeric vector}
    \item{\code{x101}}{a numeric vector}
    \item{\code{x102}}{a numeric vector}
    \item{\code{x103}}{a numeric vector}
    \item{\code{x104}}{a numeric vector}
    \item{\code{x105}}{a numeric vector}
    \item{\code{x106}}{a numeric vector}
    \item{\code{x107}}{a numeric vector}
    \item{\code{x108}}{a numeric vector}
    \item{\code{x109}}{a numeric vector}
    \item{\code{x110}}{a numeric vector}
    \item{\code{x111}}{a numeric vector}
    \item{\code{x112}}{a numeric vector}
    \item{\code{x113}}{a numeric vector}
    \item{\code{x114}}{a numeric vector}
    \item{\code{x115}}{a numeric vector}
    \item{\code{x116}}{a numeric vector}
    \item{\code{x117}}{a numeric vector}
    \item{\code{x118}}{a numeric vector}
    \item{\code{x119}}{a numeric vector}
    \item{\code{x120}}{a numeric vector}
    \item{\code{x121}}{a numeric vector}
    \item{\code{x122}}{a numeric vector}
    \item{\code{x123}}{a numeric vector}
    \item{\code{x124}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\examples{
data(dataR)
## maybe str(dataR) ; plot(dataR) ...
}
\keyword{datasets}
